Remoed dashboard and renamed home to events since we have landing page now.
