+++ code/src/App.css
@@ -23,7 +23,6 @@
   overflow-x: hidden;
 }
 
-.dashboard-page,
 .admin-page,
 .organizations-page {
   min-height: 100vh;
